use separate templates for units and other channel types
